spaces/WebPTalkHead/requirements.txt
numpy
torch
pillow
spaces
webp